doc: reference old release notes in archived release docs
Instead of trying to maintain the release notes for old releases in the
current release, reference them in a previous archive.  This eliminates
the need to fix doc build errors by editing old release notes that
reference material that no longer exists in the newer release.

We use the intersphinx extension to support references to
Sphinx-generated content from other projects.  In our case, we're
referencing the archived version of v2.7 documents that was the last
release to generate release notes for all previous releases. This import
is done in conf.py.

(Alternatively we could reference the release notes for a release in
that specific release's archive, but that means fetchind the intersphinx
inventory file (objects.inv) for every release, and seems like
overkill.)

This modification still allows putting multiple release notes in the
doc/release_notes folder and they'll show up in the release notes index
as they have in the past.  Update the separate "archived" release notes
section manually as needed.

Also, adds a draft of the v3.0 release notes to verify the new release
notes index page layout.

Fixes: #7171

Tracked-On: #7171

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>