Provides a Git pre-commit hook to check Intel copyright header year
Over time, the copyright year or year range becomes out of date if not
kept updated as changes to files with an Intel copyright header are
edited. This client-side pre-commit hook reminds the contributor to
update the year in the copyright header if the contributor has an
intel.com email domain.

Recommended usage is to create a symbolic link to this file:

   cd acrn-hypervisor/.git/hooks; ln -s ../.pre-commit-intel pre-commit

If needed, the check can be ignored by using   git commit --no-verify

Tracked-On: #7668

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>