doc: update vUART tutorials
The current code assign COM2 to S5 feature as a commutation vUART, so
the COM2's IO port which value is 0x2F8 could not assign to another
vUART connection.

This patch add a note and modify the legacy vUART sample, change the
default IO port to 0x3E8 which was known as COM3.

Tracked-On: #6690
Signed-off-by: Chenli Wei <chenli.wei@linux.intel.com>